# docs/swagger.yaml
definitions:
...
  model.Attribute:
    properties:
      characterID:
        type: integer
      charisma:
        type: integer
      constitution:
        type: integer
      dexterity:
        type: integer
      id:
        type: integer
      intelligence:
        type: integer
      strength:
        type: integer
      wisdom:
        type: integer
    type: object
...
  model.Character:
    properties:
      alias:
        type: string
      ancestryID:
        type: integer
      attribute:
        $ref: '#/definitions/model.Attribute'
      backgroundID:
        type: integer
      boost:
        $ref: '#/definitions/model.CharacterBoost'
      characterClassID:
        type: integer
      characterDefence:
        $ref: '#/definitions/model.CharacterDefence'
      characterFeat:
        $ref: '#/definitions/model.CharacterFeat'
      characterInfo:
        $ref: '#/definitions/model.CharacterInfo'
      characterItem:
        items:
          $ref: '#/definitions/model.CharacterItem'
        type: array
      characterSkill:
        items:
          $ref: '#/definitions/model.CharacterSkill'
        type: array
      characterSpell:
        items:
          $ref: '#/definitions/model.CharacterSpell'
        type: array
      id:
        type: integer
      last_name:
        type: string
      level:
        type: integer
      name:
        type: string
      raceID:
        type: integer
      slot:
        items:
          $ref: '#/definitions/model.Slot'
        type: array
      userID:
        type: integer
    type: object
  model.CharacterBoost:
    properties:
      ancestryBoost:
        type: integer
      backgroundBoost:
        type: boolean
      characterID:
        type: integer
      classBoost:
        type: boolean
      freeBoost:
        type: integer
      id:
        type: integer
    type: object
...
  model.CharacterDefence:
    properties:
      armorClass:
        type: integer
      characterID:
        type: integer
      dying:
        type: integer
      fortitude:
        $ref: '#/definitions/model.MasteryLevel'
      heavyArmor:
        $ref: '#/definitions/model.MasteryLevel'
      hitPoint:
        type: integer
      id:
        type: integer
      lightArmor:
        $ref: '#/definitions/model.MasteryLevel'
      maxHitPoint:
        type: integer
      mediumArmor:
        $ref: '#/definitions/model.MasteryLevel'
      perception:
        $ref: '#/definitions/model.MasteryLevel'
      reflex:
        $ref: '#/definitions/model.MasteryLevel'
      speed:
        type: integer
      temporaryHitPoint:
        type: integer
      unarmed:
        $ref: '#/definitions/model.MasteryLevel'
      will:
        $ref: '#/definitions/model.MasteryLevel'
      wounded:
        type: boolean
    type: object
...
  model.CharacterFeat:
    properties:
      characterID:
        type: integer
      featID:
        type: integer
      id:
        type: integer
    type: object
...
  model.CharacterInfo:
    properties:
      bulk:
        type: number
      characterID:
        type: integer
      classDC:
        type: integer
      heroPoint:
        type: integer
      id:
        type: integer
      maxBulk:
        type: number
    type: object
  model.CharacterItem:
    properties:
      armor:
        items:
          $ref: '#/definitions/model.Slot'
        type: array
      character:
        $ref: '#/definitions/model.Character'
      characterID:
        type: integer
      firstWeapon:
        items:
          $ref: '#/definitions/model.Slot'
        type: array
      id:
        type: integer
      item:
        $ref: '#/definitions/model.Item'
      itemID:
        type: integer
      quantity:
        type: integer
      secondWeapon:
        items:
          $ref: '#/definitions/model.Slot'
        type: array
    type: object
...
  model.CharacterSkill:
    properties:
      characterID:
        type: integer
      id:
        type: integer
      mastery:
        $ref: '#/definitions/model.MasteryLevel'
      skillID:
        type: integer
    type: object
...
  model.CharacterSpell:
    properties:
      characterID:
        type: integer
      id:
        type: integer
      spellID:
        type: integer
    type: object
...
  model.CreateUser:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  model.Item:
    properties:
      bulk:
        type: number
      characterItem:
        items:
          $ref: '#/definitions/model.CharacterItem'
        type: array
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      ownerID:
        type: integer
      ownerType:
        type: string
      price:
        type: string
    type: object
...
  model.MasteryLevel:
    enum:
    - None
    - Train
    - Expert
    - Master
    - Legend
    type: string
    x-enum-varnames:
    - None
    - Train
    - Expert
    - Master
    - Legend
...
  model.Slot:
    properties:
      armor:
        $ref: '#/definitions/model.CharacterItem'
      armorID:
        type: integer
      character:
        $ref: '#/definitions/model.Character'
      characterID:
        type: integer
      firstWeapon:
        $ref: '#/definitions/model.CharacterItem'
      firstWeaponID:
        type: integer
      id:
        type: integer
      secondWeapon:
        $ref: '#/definitions/model.CharacterItem'
      secondWeaponID:
        type: integer
    type: object
...
  model.UserExternal:
    properties:
      admin:
        type: boolean
      characters:
        items:
          $ref: '#/definitions/model.Character'
        type: array
      id:
        type: integer
      username:
        type: string
    required:
    - username
    type: object
...
info:
  contact: {}
  description: Api for Pet Project
  title: Kingdom Api
  version: 0.1.0
paths:
  /a/rabbit:
    post:
      consumes:
      - application/json
      description: Returns all users
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: user details
          schema:
            $ref: '#/definitions/model.UserExternal'
      summary: Returns all users
      tags:
      - A
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: JWT security accessToken. Please add it in the format "Bearer {AccessToken}"
      to authorize your requests.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"